데이터 타입 변경
float 타입으로 변경

+++ Inc/main.h
@@ -42,7 +42,7 @@ extern "C" {
 
 /* Exported types ------------------------------------------------------------*/
 /* USER CODE BEGIN ET */
-extern uint8_t ledPos;
+extern float ledPos;
 extern uint8_t ledPosUser;
 extern uint8_t led_control_mode;
 extern uint8_t auto_time_off_mode;
